Desktop: extract ConfigScreen search helper
