+++ packages/app-desktop/gui/ConfigScreen/ConfigScreen.search.test.ts
@@ -1,20 +1,7 @@
-import { AppType, SettingItem } from '@joplin/lib/models/Setting';
+import { SettingItem } from '@joplin/lib/models/Setting';
+import { settingMatchesSearch } from './configScreenUtils';
 
 type SearchableSetting = Pick<SettingItem, 'label' | 'description'>;
-
-const settingMatchesSearch = (md: SearchableSetting, searchQuery: string): boolean => {
-	if (!searchQuery) return true;
-
-	const q = searchQuery.trim().toLowerCase();
-	if (!q) return true;
-
-	const labelText = md.label ? md.label() : '';
-	const descriptionText = md.description ? md.description(AppType.Desktop) : '';
-	const label = String(labelText || '').toLowerCase();
-	const description = String(descriptionText || '').toLowerCase();
-
-	return label.includes(q) || description.includes(q);
-};
 
 const makeSetting = (label: string, description: string): SettingItem => {
 	const setting: SearchableSetting = {
